+++ app/src/main/java/com/zhan/budget/Model/PieDataCostInterface.java
@@ -4,7 +4,7 @@ package com.zhan.budget.Model;
  * Created by Zhan on 16-06-06.
  */
 public interface PieDataCostInterface {
-    float getPieDataCost();
+    double getPieDataCost();
 
     String getPieDataName();
